Refactor tools system in order to be able to bundle them separately (bundle.js was now 2MB+ minified...)
ToolManager is now a separate component.
Tools are no longer modules, but included directly from the main html file, and register themselves to the ToolManager.
draco_decoder is no longer directly included, so that it does not increase the size of the main bundle and is only loaded on demand.

factor-bundle+exorcist are used in addition to browserify to be able to handle multiple bundles and extract source maps for production/analysis.
source-map-explorer can be used to provide info on the content of each bundle:

  for f in bundle.*.js; do echo $f; ./node_modules/.bin/source-map-explorer $f $f.map ; done