Refactor: browser-side main scripts to use BaseToolManager and GLTFContainer.
Viewer is no longer responsible for constructing the dataset used by the tools.
This allows for much more decoupling (tools can be used without a viewer, and reciprocically).
This refactor comes at the cost of currently requesting all resources twice (but the browser should cache the data).
The goal for the tools is to make them usable in more contexts (such as server-side, or as command line tools)
The goal for the Viewer (and GLTFLoader) is to minimize the changes compared to upstream versions.