[tools] Add ToolPackZIP allowing to save the modified asset as a ZIP archive
The main challenge is to unpack the GLB body if it exists.
I chose to split up any image or draco mesh it contains, as these files can be useful independently of the main gltf file.

When the GLB was generated by ToolPackGLB, the image file names are preserved (as ToolPackGLB copy the original uri into the name property).