[draco] refactor the design of DRACOLoader to provide the attributes data as typed arrays instead of creating a THREE.BufferGeometry directly.
This allows to apply the Draco decompression pass at the end of GLTFLoader.loadAccessors(), instead of within the primitives conversion to THREE.BufferGeometry in GLTFLoader.loadPrimitives (bypassing/copying all the standard primitive creation code).

This allows to support meshes mixing compressed and uncompressed attributes (which is required by the spec proposal), but requires that the "data-less" compressed accessors ids are unique between different compressed meshes (which may or may not be required by the spec, but is already often required due to the attribute-specific information that must be preserved such as min/max).

Also, the decoded attributes are decoded into the specific type of array (Float32Array, Uint32Array, Uint16Array, ...), according to the accessor's componentType, so that the decompression pass produces exactly the same accessors arrays as if they were provided normally.