[draco] FIX decoding of int attributes encoded as floats
The current Draco encoder javascript API only supports Float
attributes. As a result, when int values are encoded (such as for
skinning joints ids), due to quantization they may be decoded into
float values that are slightly below the input values. In this case
GetAttributeIntForAllPoints() does a simple casting that will
result in integers that can be reduced by 1. Instead, until either
the encoder or the decoder API is fixed, we always use
GetAttributeFloatForAllPoints(), and convert to integers by
rounding to the nearest value with Math.round().

A good illustration of this problem is when loading the animated
CesiumMan sample (as encoded with ToolDracoCompressor).