[draco] EXPERIMENTAL to-be-discussed support for morph targets, both in encoder and decoder.
This is actually quite straightforward (after the decoder refactoring), the targets array
is copied into the draco extension exactly as for the attributes, i.e. keys have the same
meaning (morph target index / attribute name), but accessors id values are
replaced by draco attribute ids.

DO NOT USE IN PRODUCTION, until this is either accepted or rejected as part of the official extension.